# server/openapi.spec.yaml
openapi: 3.0.2
info:
  title: DR.CODE Project
  version: 0.0.1
  description: This is the API documentation for DR.CODE Kecskemét's management application
servers:
  - url: http://127.0.0.1:3000
    description: localhost
paths:
...
  /location/get:
    get:
      summary: Get locations by id
      description: Get one or more location(s) with the ability to fetch record links.
      tags:
        - location
      security:
        - BearerJwt: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                title: GetLocationsByIdOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Location"
          description: The request was successful and the requested locations were retrieved.
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters:
        - $ref: "#/components/parameters/ids"
      operationId: getLocationsById
...
  /location/update:
    post:
      summary: Update a location
      description: This endpoint can be used to update a location as an administrator.
      tags:
        - location
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful, and the location was updated.
          content:
            application/json:
              schema:
                type: object
                title: UpdateALocationOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - location
                    properties:
                      location:
                        $ref: "#/components/schemas/Location"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: updateALocation
      requestBody:
        description: The request body for updating a location.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ID"
                name:
                  type: string
                  example: Kecskeméti Kodály Zoltán Ének-Zenei Általános Iskola
                address:
                  type: string
                  example: 6000 Kecskemét, Dózsa György út 22.
                contact_email:
                  $ref: "#/components/schemas/Email"
                contact_phone:
                  $ref: "#/components/schemas/PhoneNumber"
  /location/create:
    post:
      summary: Create a new location
      description: This endpoint is used to create a new location. Only available to admins.
      tags:
        - location
      security:
        - BearerJwt: []
      responses:
        "201":
          description: The request was successful, and the location was created.
          content:
            application/json:
              schema:
                type: object
                title: CreateALocationCreated
                required:
                  - code
                  - data
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - location
                    properties:
                      location:
                        $ref: "#/components/schemas/Location"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: createALocation
      requestBody:
        description: The request body for creating a new location.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - address
                - contact_email
                - contact_phone
              properties:
                name:
                  type: string
                  example: Kecskeméti Kodály Zoltán Ének-Zenei Általános Iskola
                address:
                  type: string
                  example: 6000 Kecskemét, Dózsa György út 22.
                contact_email:
                  $ref: "#/components/schemas/Email"
                contact_phone:
                  $ref: "#/components/schemas/PhoneNumber"
  # # # # # # # # # #
  # SUBJECT  ROUTES #
  # # # # # # # # # #
...
  /subject/get:
    get:
      summary: Get subjects by id
      description: Get one or more subject(s) with the ability to fetch record links.
      tags:
        - subject
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful and the requested subjects were retrieved.
          content:
            application/json:
              schema:
                type: object
                title: GetSubjectsByIdOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subject"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters:
        - $ref: "#/components/parameters/ids"
      operationId: getSubjectsById
...
  /subject/update:
    post:
      summary: Update a subject
      description: This endpoint can be used to update a subject as an administrator.
      tags:
        - subject
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful, and the subject was updated.
          content:
            application/json:
              schema:
                type: object
                title: UpdateASubjectOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - subject
                    properties:
                      subject:
                        $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: updateASubject
      requestBody:
        description: The request body for updating a subject.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ID"
                name:
                  type: string
                  example: Scratch 1
                  description: The name of the subject
  /subject/create:
    post:
      summary: Create a new subject
      description: This endpoint is used to create a new subject. Only available to admins.
      tags:
        - subject
      security:
        - BearerJwt: []
      responses:
        "201":
          description: The request was successful, and the subject was created.
          content:
            application/json:
              schema:
                type: object
                title: CreateASubjectCreated
                required:
                  - code
                  - data
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - subject
                    properties:
                      subject:
                        $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: createASubject
      requestBody:
        description: The request body for creating a new subject.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Scratch 1
                  description: The name of the subject
  # # # # # # # # # #
  # STUDENT  ROUTES #
  # # # # # # # # # #
...
  /student/get:
    get:
      summary: Get students by id
      description: Get one or more student(s) with the ability to fetch record links.
      tags:
        - student
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful and the requested students were retrieved.
          content:
            application/json:
              schema:
                type: object
                title: GetStudentsByIdOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters:
        - $ref: "#/components/parameters/ids"
      operationId: getStudentsById
...
  /student/update:
    post:
      summary: Update a student
      description: This endpoint can be used to update a student as an administrator.
      tags:
        - student
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful, and the student was updated.
          content:
            application/json:
              schema:
                type: object
                title: UpdateAStudentOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - student
                    properties:
                      student:
                        $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: updateAStudent
      requestBody:
        description: The request body for updating a student.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ID"
                name:
                  type: string
                  example: John Doe
                parent:
                  type: object
                  description: The name of the parent or guardian
                  required:
                    - name
                    - email
                    - phone
                  properties:
                    name:
                      type: string
                      example: Jane Doe
                    email:
                      $ref: "#/components/schemas/Email"
                    phone:
                      $ref: "#/components/schemas/PhoneNumber"
                email:
                  $ref: "#/components/schemas/Email"
                phone:
                  $ref: "#/components/schemas/PhoneNumber"
  /student/create:
    post:
      summary: Create a new student
      description: This endpoint is used to create a new student. Only available to admins.
      tags:
        - student
      security:
        - BearerJwt: []
      responses:
        "201":
          description: The request was successful, and the student was created.
          content:
            application/json:
              schema:
                type: object
                title: CreateAStudentCreated
                required:
                  - code
                  - data
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - student
                    properties:
                      student:
                        $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: createAStudent
      requestBody:
        description: The request body for creating a new student.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - parent
                - email
                - phone
              properties:
                name:
                  type: string
                  example: John Doe
                parent:
                  type: object
                  description: The name of the parent or guardian
                  required:
                    - name
                    - email
                    - phone
                  properties:
                    name:
                      type: string
                      example: Jane Doe
                    email:
                      $ref: "#/components/schemas/Email"
                    phone:
                      $ref: "#/components/schemas/PhoneNumber"
                email:
                  $ref: "#/components/schemas/Email"
                phone:
                  $ref: "#/components/schemas/PhoneNumber"
  # # # # # # # # #
  # LESSON ROUTES #
  # # # # # # # # #
...
components:
  schemas:
...
    Location:
      type: object
      description: A location object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created:
          $ref: "#/components/schemas/Datetime"
        name:
          type: string
          description: The name of the location
          example: Kecskeméti Kodály Zoltán Ének-Zenei Általános Iskola
        address:
          type: string
          description: The address of the location
          example: 6000 Kecskemét, Dózsa György út 22.
        contact_email:
          $ref: "#/components/schemas/Email"
        contact_phone:
          $ref: "#/components/schemas/PhoneNumber"
      required:
        - id
        - created
        - name
        - address
        - contact_email
        - contact_phone
...
    Subject:
      type: object
      description: A subject object
      required:
        - id
        - created
        - name
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created:
          $ref: "#/components/schemas/Datetime"
        name:
          type: string
          description: The name of the subject
          example: Scratch
...
    Student:
      type: object
      description: A student object
      required:
        - id
        - created
        - name
        - parent
        - email
        - phone
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created:
          $ref: "#/components/schemas/Datetime"
        name:
          type: string
          description: The name of the student
          example: John Doe
        parent:
          type: object
          description: The name of the parent or guardian
          required:
            - name
            - email
            - phone
          properties:
            name:
              type: string
              description: The name of the parent or guardian
              example: Jane Doe
            email:
              $ref: "#/components/schemas/Email"
            phone:
              $ref: "#/components/schemas/PhoneNumber"
        email:
          $ref: "#/components/schemas/Email"
        phone:
          $ref: "#/components/schemas/PhoneNumber"
...
    Email:
      type: string
      format: email
      description: A string in email address format.
    PhoneNumber:
      type: string
      format: telephone
      example: "+36301131313"
      description: A telephone number starting with country code
    ID:
      type: string
      description: A surrealdb id in the format `table:id`
      example: sometable:1234
...
    Datetime:
      type: string
      description: ISO 8601 date-time string
      format: date-time
      example: 2023-10-15T14:30:00Z
  responses:
    NotFound:
      description: The requested resource was not found on the server.
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - code
            properties:
              code:
                enum:
                  - not_found
                type: string
                example: not_found
              message:
                type: string
                example: Some error message...
    BadRequest:
      description: Something is missing or incorrect in the client's request, so the server cannot process it.
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - code
            properties:
              code:
                enum:
                  - server_error
                  - bad_request
                  - unauthorized
                  - invalid_credentials
                  - invalid_request_body
                  - fields_required
                  - fields_invalid
                  - password_too_weak
                  - email_in_use
                  - invalid_invite
                type: string
                example: bad_request
              message:
                type: string
                example: Some error message...
    Unauthorized:
      description: Authentication is required to access the requested resource. The client must include the appropriate credentials.
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - code
            properties:
              code:
                enum: 
                  - unauthorized
                type: string
                example: unauthorized
              message:
                type: string
                example: You are not authorized to perform this action
  parameters: 
    ids:
      name: ids
      description: Ids in surrealdb format (table:id) separated by commas
      in: query
      style: form
      schema:
        $ref: "#/components/schemas/ID"
      explode: true
      required: true
      allowReserved: false
  securitySchemes:
    BearerJwt:
      type: http
      scheme: bearer
      description: This is the token used for authenticating users
      bearerFormat: Json Web Token (JWT)
...